Boot argument files for NAND boot.
Adding the env.txt file that was being created manually for NAND
boots.  This can now be used as a communication path between the
teams on bootargs.

Patch Set 2:
Made changes suggested by Anu.  I believe the double quotes are
correct around the 0:4M.

Patch Set 3:
This time I really made the changes suggested by Anu.  Also added
automatic copy to the root directory.  This will ensure that
the env.txt appears in the right place for everyone to create a
filesystem.