PM: ZOOM3: Enabling SmartReflex through sysfs
Enabling SmartReflex for Zoom3 OMAP3630 using a
sysfs interface. This change is not required
for Zoom2 OMAP3430 since it is enabled by default
in the kernel.

By doing this we won't affect the ZOOM2 but we'll
get the following messages regarding the SR status:

"SR1: VDD autocomp is already active"
"SR2: VDD autocomp is already active"

Once, we get the proper characteristics values for
OMAP3630, SR will be enable by default in the kernel
and this patch must be reverted.

Signed-off-by: Leed Aguilar <leed.aguilar@ti.com>