USB: ADB: Remove 'setprop service.adb.tcp.port' from init.rc
This patch removes the setprop service.adb.tcp.port setting from
the init.rc files.

With the new code changes in the ADB code from google, if this
property is set, ADB daemon will always listen to TCP port and
ADB over USB does not work. So inorder for ADB over USB to work
out of the box, remove this setprop from the init.rc file.

Inorder to make ADB over ethernet work:

On the OMAP board execute:
$ setprop service.adb.tcp.port 5555
$ stop adbd
$ start adbd

On the Host machine execute:
$ export ADBHOST=<Board-IP>
$ ./adb kill-server
$ ./adb start-server

Change-Id: Ia692e87ce25027dfad62f7e3b6566d2d4f35fe36
Signed-off-by: Moiz Sonasath <m-sonasath@ti.com>