Changed TARGET_CPU_ABI from armeabi to armeabi-v7a
Issue:
The Flash APK that is built for ARM7 does not install on Froyo. The Package Manager rejects it. The Flash APK must be built for ARM7
since it utilizes some neon routines. If the Flash APK is built for ARM5, the Package Manager has no problem installing it.

Changing the TARGET_CPU_ABI from armeabi to armeabi-v7a  helps resolve the problem.

Change-Id: Ie346df95741ad728a6fd8804f43d406a5bf72aae
Signed-off-by: Anu Sundararajan <sanuradha@ti.com>